requirements.txt
anyio==3.6.2
azure-core==1.26.4
azure-storage-blob==12.16.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
databases==0.6.2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
isodate==0.6.1
ormar==0.12.1
pycparser==2.21
pydantic==1.10.4
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
